changes made as requested: 1. changed the description of Destination in ILineEvents 2. deleted files LineEvents.cs and LineEventEnum.cs 3. fixed the empty collection returned.
