--- src/LineBot/Events/LineEventEnum.cs
@@ -1,67 +1,0 @@
-﻿// Copyright 2017-2018 Dirk Lemstra (https://github.com/dlemstra/line-bot-sdk-dotnet)
-//
-// Dirk Lemstra licenses this file to you under the Apache License,
-// version 2.0 (the "License"); you may not use this file except in compliance
-// with the License. You may obtain a copy of the License at:
-//
-//   https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations
-// under the License.
-
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace Line
-{
-    internal class LineEventEnum : IEnumerator
-    {
-        private ILineEvent[] _list;
-
-        private int _position = -1;
-
-        public LineEventEnum(ILineEvent[] lineEvents)
-        {
-            _list = lineEvents;
-        }
-
-        object IEnumerator.Current
-        {
-            get
-            {
-                return Current;
-            }
-        }
-
-        public ILineEvent Current
-        {
-            get
-            {
-                try
-                {
-                    return _list[_position];
-                }
-                catch (IndexOutOfRangeException)
-                {
-                    throw new InvalidOperationException();
-                }
-            }
-        }
-
-        public bool MoveNext()
-        {
-            _position++;
-            return _position < _list.Length;
-        }
-
-        public void Reset()
-        {
-            _position = -1;
-        }
-    }
-}
